Encrypts the message using the parameter of three.
I had to define a variable for the parameter, and just for now, I
initiated it with the value of three. Using this parameter, we loop
through the message, raising each and every character in value. If the
character in question becomes larger than the alphabet, we loop it around.

I had forgotten to uppercase the message before encryption, so much of the
encryption was messed up, since all of my substitutions were meant to
operate only on uppercase letters.