Appended some really crude code for userprompting.
I have yet to manage empty or multiple inputage, and I also want to allow
the users to utilize both encrypting and decrypting, but for now, this
will be enough to at least begin implementing some of my other functions.

And maybe we should implement some of this as an applet. :]